misc: rename files according to extensions guidelines
