+++ src/components/Inputs/MillisecondInput/internal/Millisecond.types.ts
